refactor(web): reorganize utility imports and update constants for team members
- Moved utility functions for team members and task updates from '@/core/utils' to '@/core/lib/utils' for better structure.
- Updated the ITEMS_LENGTH_TO_VIRTUALIZED constant from 25 to 8 to optimize rendering performance.
- Cleaned up imports across various components to reflect the new utility paths.
