+++ apps/web/core/lib/utils/task.utils.ts
@@ -1,3 +1,3 @@
-import { Queue } from '../lib/utils';
+import { Queue } from '.';
 
 export const taskUpdateQueue = new Queue(1);
