+++ apps/web/core/stores/tasks/task-filter.ts
@@ -1,4 +1,4 @@
-import { TeamMemberFilterType } from '@/core/utils/team-members.utils';
+import { TeamMemberFilterType } from '@/core/lib/utils/team-members.utils';
 import { atom } from 'jotai';
 
 export const taskBlockFilterState = atom<TeamMemberFilterType>('all');
